requirements.txt
streamlit==1.37.0


langchain==0.2.11
langchain-core==0.2.32
langchain-community==0.2.10
langchain-openai==0.1.8
langchain-text-splitters==0.2.2


openai==1.109.1

faiss-cpu==1.7.4
python-dotenv==1.0.1
tqdm==4.67.1
